Commentaire + Fix

+++ Lieu.py
@@ -6,5 +6,6 @@ class Lieu:
         self.x = x
         self.y = y
 
-    def distance(self, x2, y2) -> int:
+    # Renvoie la distance euclidienne entre 2 points.
+    def distance(self, x2, y2) -> float:
         return sqrt((self.x - x2)**2 + (self.y - y2)**2)
